client: Flesh out some more of the desired code-flow and add comments about it
This copies a lot of the flow and logic from the Python code base directly into
functions, and while all arguments etc. are place-holders for now, this is
starting to look a bit more like the expected flow for the general application.

With a healthy amount of comments and blocks, it should make diffs a lot less
cumbersome to work with in the future.