cert/subjects: Implement our subject generation code
This implements both work-around subjects and creating new and untouched
subjects.

Comparing subjects is implemented as string wrapping functions, sadly due to
how difficult it was to compare them as objects.

Replacement is done as a pass-by-text rather than copying data objects, but
that matches how the old code does it, and shouldn't be a big concern as it
also ensures that we are utf8-clean.