caramel client: Move all file IO into the command line tools
This moves all file IO into caramel-client and out of certs and network.rs

This caused some of the API's to change (filename => vector, etc.), which then
allowed me to further restructure the certs mod.

Now most of the public api is a simple "match" wrapper around a similar
"private" function that works with/on OpenSSL data.

So verify_foobar will have a  `openssl_verify_foobar` function to match it that
takes similar data.

Issue #4