network: Move get_root_crt to use the new Error types
This splits get_root_crt into an inner ("curl_") version and an outer version,
in order to make error handling smoother.

Many of the .unwrap() calls in the file are being removed, as we can now conver
error types to our own error data type, and log them if need be.