tests: Add CSR generation tests and CSR validation tests
This adds test-data for `ca.modio.se` which is the only case where work-around
for subject is needed.

The test adds a "Golden" copy generated by our existing client, and a test to
check that the new client generates the same data.

And a bug-fix to make it generate the right data.