network: Return correct byte count from buffer copies
Our write_functions returned the amount of bytes in our _buffer_, rather than
the amount of bytes we wrote.

This is worked for _most_ cases where the transmission happened in a single
block, but when network conditions changed, this caused failures whenever we
had to cause two writes rather than one.

Caught by the integration tests going flakey.

Closes: ModioAB/caramel-client-rs#41