Dockerfile.build: Downgrade debian to stretch
This downgrades the debian container to oldstable, and re-implements the
upstream rust/cargo container at that.

The reasoning for this is to allow us build & deploy binaries on debian 9,
which due to how glibc-versioning works requires us to have an older glibc to
link against.

As the arm32-musl target isn't stable yet for rust, and we also need
gcc-implementations, it was hairier than I felt in the mood to deal with to
enable cross-compilation.

Issue: ModioAB/caramel-client-rs#55