Set file permissions for certificat and key files
The private keys are now generated with the correct file permissions.
If a private key already exists, we check if the file permission
matches the one specified in the caramel client. A warning is given
if the user and/or the group permissions differs. If the file
permission for other user differ an error is displayed and the
execution is terminated.

Issue #29