Statically link against curl and openssl
As tests were failing in CI due to a segfault in libssl1.1 and I could
not reproduce it locally, or in the same container, I went with the easy
fix of statically linking the tools in order to make sure I have the
same version of openssl.

This has some effects on binary size and may cause interesting joy with
cross compiling, so the CI test should be done for that.