Improve error messages around rejected CSR data
This improves the error messages, wrapping the network code to output
strings rather than ascii as binary, if possible, and also converts the
functions around cert to return an error rather than panicking.

End result is still an error, with a message to delete .csr files, but
it should no longer panic.